moved back to util for mdx-loader
