+++ packages/gatsby-mdx/gatsby/on-create-node.js
@@ -1,7 +1,5 @@
-const crypto = require("crypto");
 const defaultOptions = require("../utils/default-options");
-const mdx = require("../utils/mdx");
-const extractExports = require("../utils/extract-exports");
+const createMDXNode = require("../utils/create-mdx-node");
 
 module.exports = async (
   { node, loadNodeContent, actions, createNodeId },
@@ -23,40 +21,11 @@ module.exports = async (
 
   const content = await loadNodeContent(node);
 
-  const code = await mdx(content, options);
-
-  // extract all the exports
-  const { frontmatter, ...nodeExports } = extractExports(code);
-
-  const mdxNode = {
+  const mdxNode = await createMDXNode({
     id: createNodeId(`${node.id} >>> Mdx`),
-    children: [],
-    parent: node.id,
-    internal: {
-      content: content,
-      type: "Mdx"
-    }
-  };
-
-  mdxNode.frontmatter = {
-    title: ``, // always include a title
-    ...frontmatter,
-    _PARENT: node.id
-  };
-
-  mdxNode.excerpt = frontmatter.excerpt;
-  mdxNode.exports = nodeExports;
-  mdxNode.rawBody = content;
-
-  // Add path to the markdown file path
-  if (node.internal.type === `File`) {
-    mdxNode.fileAbsolutePath = node.absolutePath;
-  }
-
-  mdxNode.internal.contentDigest = crypto
-    .createHash(`md5`)
-    .update(JSON.stringify(mdxNode))
-    .digest(`hex`);
+    node,
+    content
+  });
 
   createNode(mdxNode);
   createParentChildLink({ parent: node, child: mdxNode });
